#51 separate template component on edit and read
